@@ -1,14 +1,17 @@
 package com.elyeproj.demoglide
 
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.widget.Toast
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.resource.drawable.DrawableTransitionOptions
 import com.bumptech.glide.request.RequestOptions
+import com.elyeproj.demoglide.databinding.ActivityMainBinding
 import kotlinx.android.synthetic.main.activity_main.*
 
 class MainActivity : AppCompatActivity(), MyImageRequestListener.Callback {
+    private lateinit var binding: ActivityMainBinding
+
     override fun onFailure(message: String?) {
         Toast.makeText(this, "Fail to load: $message", Toast.LENGTH_SHORT).show()
     }
@@ -19,7 +22,8 @@ class MainActivity : AppCompatActivity(), MyImageRequestListener.Callback {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
 
         loadImage("https://images.pexels.com/photos/46239/salmon-dish-food-meal-46239.jpeg?auto=compress&cs=tinysrgb&h=50",
                 "https://images.pexels.com/photos/46239/salmon-dish-food-meal-46239.jpeg?auto=compress&cs=tinysrgb&h=1000"
@@ -38,6 +42,6 @@ class MainActivity : AppCompatActivity(), MyImageRequestListener.Callback {
                         .apply(requestOption))
                 .apply(requestOption)
                 .listener(MyImageRequestListener(this))
-                .into(my_image_view)
+                .into(binding.myImageView)
     }
 }
